variadic functions task 3

+++ variadic_functions/3-print_all.c
@@ -7,7 +7,7 @@
 void print_char(va_list arg)
 {
 	char character;
-	character = va_arg(arg,int);
+	character = va_arg(arg, int);
 	printf("%c", character);
 }
 /**
@@ -23,7 +23,7 @@ void print_int(va_list arg)
 void print_float(va_list arg)
 {
 	float dec;
-	dec = va_arg(arg, int);
+	dec = va_arg(arg, double);
 	printf("%f", dec);
 }
 /**
